personSaga and personSlice renamed to searchSaga and searchSlice, store and imports adjusted
