+++ src/core/saga.js
@@ -2,13 +2,13 @@ import { all } from "redux-saga/effects";
 import { moviesSaga } from "../features/movies/moviesSaga";
 import { peopleSaga } from "../features/people/peopleSaga";
 import { paginationSaga } from "../common/Pagination/paginationSaga";
-import { personSaga } from "../features/people/personSaga";
+import { searchSaga } from "../features/people/searchSaga";
 
 export default function* rootSaga() {
     yield all ([
         moviesSaga(),
         peopleSaga(),
-        personSaga(),
+        searchSaga(),
         paginationSaga(),
     ]);
 }
