+++ src/core/store.js
@@ -2,9 +2,9 @@ import { configureStore } from "@reduxjs/toolkit";
 import peopleReducer from "../features/people/peopleSlice";
 import moviesReducer from "../features/movies/moviesSlice";
 import paginationReducer from "../common/Pagination/paginationSlice";
-import personReducer from "../features/people/personSlice";
 import createSagaMiddleware from "redux-saga";
 import rootSaga from "./saga";
+import searchReducer from "../features/people/searchSlice";
 
 const sagaMiddleware = createSagaMiddleware();
 
@@ -13,7 +13,7 @@ export default configureStore({
         people: peopleReducer,
         movies: moviesReducer,
         pagination: paginationReducer,
-        person: personReducer,
+        search: searchReducer,
     },
     middleware: (getDefaultMiddleware) =>
         getDefaultMiddleware().concat(sagaMiddleware),
