+++ src/features/people/PopularPeople/index.js
@@ -5,11 +5,10 @@ import { selectPeople } from "../peopleSlice";
 import { Pagination } from "../../../common/Pagination";
 import { useQueryParameters } from "../../../common/Navigation/Search/queryParameters";
 import searchQueryParamName from "../../../common/Navigation/Search/searchQueryParamName";
-import { selectSearchResult, selectSearchStatus } from "../personSlice";
+import { selectSearchResult, selectSearchStatus } from "../searchSlice";
 
 export const PopularPeople = () => {
     const query = useQueryParameters(searchQueryParamName);
-    const searchStatus = useSelector(selectSearchStatus);
     const people = useSelector(state =>
         query?.trim()
             ? selectSearchResult(state)
